✨ Integrate `DynamicNotchProgress` into `DynamicNotchInfoIcon`
